Refactor AvailableSpaces related services and rename files to exclude page suffix
