+++ unischedule/lib/constants/http/http_constants.dart
@@ -1,8 +1,15 @@
 class HTTPConstants {
 
-  // BACKEND END POINTS
+  // AVAILABLE SPACES
   static const String AVAILABLE_SPACES_BASE_URL = "https://available-spaces-dot-unischedule-5ee93.uc.r.appspot.com/";
+  static const String AVAILABLE_SPACES_ENDPOINT = "spaces";
+
+  // FRIENDS
   static const String FRIENDS_BASE_URL = "https://unischedule-5ee93.uc.r.appspot.com/";
+
+  // GROUPS
   static const String GROUPS_BASE_URL = "https://unischedule-5ee93.uc.r.appspot.com/";
+
+  // EVENTS
   static const String EVENTS_BASE_URL = "https://unischedule-5ee93.uc.r.appspot.com/";
 }
