+++ unischedule/lib/constants/types/type_constants.dart
@@ -13,4 +13,12 @@ class TypeConstants {
   static const String durationIdKey = 'duration';
   static const String donationAskCard = 'donationAskCard';
   static const String feedbackCard = 'feedbackCard';*/
+
+  static const String monday = 'l';
+  static const String tuesday = 'm';
+  static const String wednesday = 'i';
+  static const String thursday = 'j';
+  static const String friday = 'v';
+  static const String saturday = 's';
+  static const String sunday = 'd';
 }
