+++ unischedule/lib/models/available_spaces/available_spaces_model.dart
@@ -1,0 +1,18 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'available_spaces_model.freezed.dart';
+part 'available_spaces_model.g.dart';
+
+@freezed
+class AvailableSpacesModel with _$AvailableSpacesModel {
+  const factory AvailableSpacesModel({
+    required String building,
+    required String room,
+    required String availableFrom,
+    required String availableUntil,
+    required int minutesAvailable,
+  }) = _AvailableSpacesModel;
+
+  factory AvailableSpacesModel.fromJson(Map<String, dynamic> json) => _$AvailableSpacesModelFromJson(json);
+}
+
