+++ unischedule/lib/models/events/event_model.dart
