+++ unischedule/lib/models/friends/friend_model.dart
