+++ unischedule/lib/models/models.dart
@@ -1,1 +1,4 @@
-export 'groups_page/group_model.dart';
+export 'groups/group_model.dart';
+export 'friends/friend_model.dart';
+export 'events/event_model.dart';
+export 'available_spaces/available_spaces_model.dart';
