+++ unischedule/lib/providers/events/events_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:unischedule/models/events_page/event_model.dart';
-import 'package:unischedule/repositories/events_page/events_repository.dart';
+import 'package:unischedule/models/models.dart';
+import 'package:unischedule/repositories/repositories.dart';
 import 'package:unischedule/services/network/events_api_service.dart';
 
 final eventsApiServiceProvider = Provider<EventsApiService>((ref) {
@@ -13,8 +13,8 @@ final eventsRepositoryProvider = Provider<EventsRepository>((ref) {
 
 final eventsProvider = FutureProvider.family<List<Event>, String>((ref, userId) {
   final repository = ref.read(eventsRepositoryProvider);
-  final searchQuery = ref.watch(searchQueryProvider);
+  final searchQuery = ref.watch(eventsSearchQueryProvider);
   return repository.getEvents(userId, searchQuery);
 });
 
-final searchQueryProvider = StateProvider<String>((ref) => "");
+final eventsSearchQueryProvider = StateProvider<String>((ref) => "");
