+++ unischedule/lib/providers/friends/friends_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:unischedule/models/friends_page/friend_model.dart';
-import 'package:unischedule/repositories/friends_page/friends_page_repository.dart';
+import 'package:unischedule/models/models.dart';
+import 'package:unischedule/repositories/repositories.dart';
 import 'package:unischedule/services/network/friends_api_service.dart';
 
 // Define el proveedor del servicio API
@@ -16,9 +16,9 @@ final friendsRepositoryProvider = Provider<FriendsRepository>((ref) {
 // Define el proveedor para obtener amigos
 final friendsProvider = FutureProvider.family<List<Friend>, String>((ref, userId) {
   final repository = ref.read(friendsRepositoryProvider);
-  final searchQuery = ref.watch(searchQueryProvider);
+  final searchQuery = ref.watch(friendsSearchQueryProvider);
   return repository.getFriends(userId, searchQuery);
 });
 
 // Proveedor para controlar el estado de la consulta de búsqueda
-final searchQueryProvider = StateProvider<String>((ref) => "");
+final friendsSearchQueryProvider = StateProvider<String>((ref) => "");
