+++ unischedule/lib/providers/providers.dart
@@ -1,0 +1,5 @@
+export 'available_spaces/available_spaces_provider.dart';
+export 'events/events_provider.dart';
+export 'friends/friends_provider.dart';
+export 'groups/groups_provider.dart';
+export 'navigation/navigation_bar_controller.dart';
