+++ unischedule/lib/repositories/repositories.dart
@@ -1,0 +1,4 @@
+export 'available_spaces/available_spaces_repository.dart';
+export 'events/events_repository.dart';
+export 'friends/friends_repository.dart';
+export 'groups/groups_repository.dart';
