+++ unischedule/lib/views/calendar_page/calendar_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:unischedule/views/calendar_page/widgets/calendar_event.dart';
 
-import '../../../providers/events_page/events_state_notifier.dart';
+import 'package:unischedule/providers/events/events_state_notifier.dart';
 
 import 'widgets/calendar_background.dart';
 import 'widgets/calendar_time_lines.dart';
