+++ unischedule/lib/views/create-class_page/widgets/color_picker_button.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:unischedule/constants/constants.dart';
+import 'package:flutter_colorpicker/flutter_colorpicker.dart';
+
+class ColorPickerButton extends StatefulWidget {
+  const ColorPickerButton({Key? key}) : super(key: key);
+
+  @override
+  _ColorPickerButtonState createState() => _ColorPickerButtonState();
+}
+
+class _ColorPickerButtonState extends State<ColorPickerButton> {
+
+  Color currentColor = ColorConstants.limerick;
+  void changeColor(Color color) {
+    setState(() => currentColor = color);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+        onPressed: () {
+          showDialog(
+            context: context,
+            builder: (BuildContext context) {
+              return AlertDialog(
+                titlePadding: const EdgeInsets.all(0.0),
+                contentPadding: const EdgeInsets.all(0.0),
+                content: SingleChildScrollView(
+                  child: ColorPicker(
+                    pickerColor: currentColor,
+                    onColorChanged: changeColor,
+                    colorPickerWidth: 300.0,
+                    pickerAreaHeightPercent: 0.7,
+                    displayThumbColor: true,
+                    enableAlpha: false,
+                    paletteType: PaletteType.hsv,
+                    pickerAreaBorderRadius:
+                    const BorderRadius.all(Radius.circular(16)),
+                  ),
+                ),
+              );
+            },
+          );
+        },
+        style: ButtonStyle(
+          backgroundColor: MaterialStateProperty.all(currentColor),
+          shape: MaterialStateProperty.all(
+            RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(4),
+            ),
+          ),
+        ),
+        child: const SizedBox(width: 24, height: 24));
+  }
+}
