+++ unischedule/lib/views/friends_page/friends_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:unischedule/models/friends_page/friend_model.dart';
+import 'package:unischedule/models/models.dart';
 import '../../../providers/friends/friends_state_notifier.dart';
 import 'package:cached_network_image/cached_network_image.dart';
